mage-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
squareup==37.1.1.20240718
gunicorn==22.0.0
python-dotenv==1.0.1
